pyproject.toml
[project]
name = "qbittorrentbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.24.0",
    "apscheduler>=3.11.2",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
    "qbittorrent-api>=2025.11.1",
    "redis>=7.1.0",
]
